+++ packages/terra-progress-bar/src/terra-dev-site/doc/progress-bar/ProgressBar.1.doc.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
+import Markdown from 'terra-markdown';
 import ReadMe from '../../../../docs/README.md';
+import ProgressBarColorDocs from '../../../../docs/progress-bar-color.md';
+import ProgressBarTwoColorsDocs from '../../../../docs/progress-bar-two-colors.md';
+import ProgressBarCustomizedColorsDocs from '../../../../docs/progress-bar-customized-colors.md';
 import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
@@ -14,8 +18,10 @@ import ProgressBarSize from '../example/ProgressBarSize';
 import ProgressBarSizeSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProgressBarSize';
 import ProgressBarColor from '../example/ProgressBarColor';
 import ProgressBarColorSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProgressBarColor';
-import ProgressBarGradient from '../example/ProgressBarGradient';
-import ProgressBarGradientSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProgressBarGradient';
+import ProgressBarTwoColors from '../example/ProgressBarTwoColors';
+import ProgressBarTwoColorsSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProgressBarTwoColors';
+import ProgressBarCustomizedColors from '../example/ProgressBarCustomizedColors';
+import ProgressBarCustomizedColorsSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProgressBarCustomizedColors';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -36,13 +42,21 @@ const DocPage = () => (
       },
       {
         title: 'Color',
+        description: <Markdown id="progress-bar-color-docs" src={ProgressBarColorDocs} />,
         example: <ProgressBarColor />,
         source: ProgressBarColorSrc,
       },
       {
-        title: 'Gradient',
-        example: <ProgressBarGradient />,
-        source: ProgressBarGradientSrc,
+        title: 'Foreground and Background Color',
+        description: <Markdown id="progress-bar-two-colors-docs" src={ProgressBarTwoColorsDocs} />,
+        example: <ProgressBarTwoColors />,
+        source: ProgressBarTwoColorsSrc,
+      },
+      {
+        title: 'Customized',
+        description: <Markdown id="progress-bar-customized-colors-docs" src={ProgressBarCustomizedColorsDocs} />,
+        example: <ProgressBarCustomizedColors />,
+        source: ProgressBarCustomizedColorsSrc,
       },
     ]}
     propsTables={[